app/requirements.txt
fastapi[all]
docker
pyyaml
uvicorn
